Add an error logging system
The `logger` package is implemented with the `CreateLogger` function,
which returns a specific logger from the `zap` package depending on
the application startup mode (dev or prod).

Rewrote the function of loading the application config from `config`
package. Now, depending on the mode of operation of the application,
the config is loaded either from the `config.yaml` file (dev),
or from the environment variables (prod).

Made a small change to the structure of the project.