Add an implementation of the RPC server for the application
Implemented the `services` package responsible for the work of the warehouse management service, contains a `SaveWarehouses` method for adding new warehouses(s).

Added to the `app` package the functions of launching the application and creating a new handler for the service.

Added the application host and port settings to `config.yml`.

Added server settings (host, port) processing to the `config` package.

Implemented warehouse validation function to the `models` package.

Added tests:
 - (UNIT) testing of warehouse model validation
 - (API) testing adding a new warehouse(s)

Added a command to run tests in the `Makefile`.

Updated the project dependencies.

Did some refactoring.