Implement a placement entity
Added a new entity to the project responsible for placing products in
the warehouse. To do this, I've expanded the functionality of the `models`
and `repositories` packages`. Added a test to check the validation of
the placement model.

Added methods for getting the product and warehouse by ID.

Added a migration file to create a placement table.

Did some refactoring.